move make_slice_1d, make_slice_2d, get_frame_indices
